requirements.txt
requests
mysql-connector-python
pandas
matplotlib
seaborn
plotly
